+++ display/display.go
@@ -172,9 +172,12 @@ func Init() {
 					// It's a full URL or search term
 					// Detect if it's a search or URL
 					if strings.Contains(query, " ") || (!strings.Contains(query, "//") && !strings.Contains(query, ".") && !strings.HasPrefix(query, "about:")) {
-						URL(viper.GetString("a-general.search") + "?" + pathEscape(query))
+						u := viper.GetString("a-general.search") + "?" + pathEscape(query)
+						cache.Remove(u) // Don't use the cached version of the search
+						URL(u)
 					} else {
 						// Full URL
+						cache.Remove(query) // Don't use cached version for manually entered URL
 						URL(query)
 					}
 					return